LG-10124: remove subclassing, since we are ok with code duplications.
